requirements.txt
llama-index-multi-modal-llms-gemini
llama-index-vector-stores-qdrant
llama-index-embeddings-gemini
llama-index-llms-gemini
llama-index==0.11.15
llama-index-readers-file
llama-index-llms-openai
llama-index-llms-anthropic
llama-index-llms-cohere
llama-index-llms-azure-openai
google-generativeai>=0.3.0 
matplotlib 
qdrant_client
python-dotenv
streamlit
llama-index-llms-nvidia
llama-index-embeddings-nvidia
markdownify
rapidocr-onnxruntime
pymupdf4llm
streamlit-authenticator
detectaicore
streamlit_pdf_viewer==0.0.12
fastapi
ipython
orjson
pymongo
motor
redis
celery
openai>=1.12.0
anthropic>=0.18.1
azure-openai>=1.0.0